--- python/primihub/client/examples/README.md
@@ -1,11 +1,0 @@
-> English | [中文](README_CN.md)
-
-Here are some examples that using Python SDK to run MPC tasks:
-
-## Prerequisite
-Please install the Python libraries according to [this document](https://docs.primihub.com/en/docs/advance-usage/python-sdk/install). Then start the PrimiHub backend.
-
-## Get started
-| Usage | Description |
-|---|---|
-| `python demo.py [primihub-ip:port]` | This a pretty simple example without actual use cases. |
