--- python/primihub/client/examples/README_CN.md
@@ -1,11 +1,0 @@
-> 中文 | [English](README.md)
-
-当前目录下有一些使用 Python SDK 执行 MPC 任务的示例代码：
-
-## 前置条件
-请先根据[这里的文档](https://docs.primihub.com/en/docs/advance-usage/python-sdk/install)安装好相应的 Python 依赖库，并启动 PrimiHub 后端服务。
-
-## 快速开始
-| 使用方法 | 介绍 |
-|---|---|
-| `python demo.py [primihub-ip:port]` | 这是一个非常简单的、没有特定场景的例子。 |
